show page number without an <input>
Display the page number in a span, since we
don't want the user to be able to switch
pages using the input box.

There is some js relying on the input box
existing, so just hide it for now. I think
it's simpler to try not to edit the forked
files too much. The viewer is pretty close to
what we want as-is.